refactor: 캐릭터 설정 상수 분리 및 코드 주석 보강 (#117)
