+++ src/features/movement/index.ts
@@ -6,3 +6,6 @@ export { useMovementStore } from "./model/useMovementStore";
 
 // UI 컴포넌트 레이어
 export { TownEngine } from "./ui/TownEngine";
+
+// 설정 상수
+export { CHARACTER_CONFIG } from "./model/config";
