+++ src/features/movement/ui/TownEngine.tsx
@@ -1,3 +1,7 @@
+/**
+ * Phaser 게임 엔진을 React 환경에 마운트하고 관리하는 엔진 컨테이너
+ * 위치 복원 상태와 연동하여 게임 인스턴스의 초기화 타이밍을 제어함
+ */
 "use client";
 
 import { useMovementSync } from "@/features/movement/hooks/useMovementSync";
@@ -9,6 +13,11 @@ import { MovementOverlay } from "@/features/movement/ui/MovementOverlay";
 import * as Phaser from "phaser";
 
 import React, { useEffect, useRef } from "react";
+
+/**
+ * Phaser 게임 엔진을 React 환경에 마운트하고 관리하는 엔진 컨테이너
+ * 위치 복원 상태와 연동하여 게임 인스턴스의 초기화 타이밍을 제어함
+ */
 
 /**
  * 마을 엔진 컨테이너 컴포넌트
